pyproject.toml
[project]
name = "geoglows-mcp-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "fastmcp>=2.12.2",
    "geoglows>=2.0.2",
    "google-adk>=1.15.1",
    "httpx>=0.28.1",
    "matplotlib>=3.10.6",
    "mcp[cli]>=1.13.1",
    "pandas>=2.3.2",
    "uvicorn>=0.35.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/tools"]